updated restricted countries PR
