+++ src/app/statev3/common/queries/is-user-restricted/IsUserRestricted.hook.ts
@@ -1,14 +1,14 @@
 import { useQuery } from "@tanstack/react-query";
-import { fetchCountryData } from "./UserCountry.fetch";
 import { CountryDataKeys } from "../../query-keys/CountryDataQueryKeys";
 import { useAccount } from "wagmi";
+import { fetchIsUserRestricted } from "./IsUserRestricted.fetch";
 
-export function useCountryData() {
+export function useFetchIsUserRestricted() {
   const { isConnected } = useAccount();
 
   return useQuery({
-    queryKey: CountryDataKeys.countryData(),
-    queryFn: () => fetchCountryData(),
+    queryKey: CountryDataKeys.isUserRestricted(),
+    queryFn: () => fetchIsUserRestricted(),
     enabled: isConnected,
   });
 }
